+++ packages/eip-5792-middleware/src/hooks/getCallsStatus.ts
@@ -7,10 +7,8 @@ import type {
 import { TransactionStatus } from '@metamask/transaction-controller';
 import type { Hex } from '@metamask/utils';
 
-import { EIP5792ErrorCode, VERSION } from '../constants';
-import { GetCallsStatusCode } from '../methods/wallet-get-calls-status';
-import type { GetCallsStatusResult } from '../methods/wallet-get-calls-status';
-import type { EIP5792Messenger } from '../types';
+import { EIP5792ErrorCode, GetCallsStatusCode, VERSION } from '../constants';
+import type { EIP5792Messenger, GetCallsStatusResult } from '../types';
 
 /**
  * Retrieves the status of a transaction batch by its ID.