+++ packages/eip-5792-middleware/src/hooks/processSendCalls.ts
@@ -18,8 +18,11 @@ import {
   MessageType,
   VERSION,
 } from '../constants';
-import type { SendCalls, SendCallsResult } from '../methods/wallet-send-calls';
-import type { EIP5792Messenger } from '../types';
+import type {
+  EIP5792Messenger,
+  SendCallsPayload,
+  SendCallsResult,
+} from '../types';
 import { getAccountKeyringType } from '../utils';
 
 /**
@@ -64,7 +67,7 @@ export type ProcessSendCallsRequest = JsonRpcRequest & {
 export async function processSendCalls(
   hooks: ProcessSendCallsHooks,
   messenger: EIP5792Messenger,
-  params: SendCalls,
+  params: SendCallsPayload,
   req: ProcessSendCallsRequest,
 ): Promise<SendCallsResult> {
   const {
@@ -156,7 +159,7 @@ async function processSingleTransaction({
   networkClientId: string;
   origin?: string;
   securityAlertId: string;
-  sendCalls: SendCalls;
+  sendCalls: SendCallsPayload;
   transactions: { params: BatchTransactionParams }[];
   validateSecurity: (
     securityRequest: ValidateSecurityRequest,
@@ -229,7 +232,7 @@ async function processMultipleTransaction({
   messenger: EIP5792Messenger;
   networkClientId: string;
   origin?: string;
-  sendCalls: SendCalls;
+  sendCalls: SendCallsPayload;
   securityAlertId: string;
   transactions: { params: BatchTransactionParams }[];
   validateSecurity: (
@@ -285,7 +288,7 @@ function generateBatchId(): Hex {
  * @param sendCalls - The sendCalls request to validate.
  * @param dappChainId - The chain ID that the dApp is connected to.
  */
-function validateSingleSendCall(sendCalls: SendCalls, dappChainId: Hex) {
+function validateSingleSendCall(sendCalls: SendCallsPayload, dappChainId: Hex) {
   validateSendCallsVersion(sendCalls);
   validateCapabilities(sendCalls);
   validateDappChainId(sendCalls, dappChainId);
@@ -301,7 +304,7 @@ function validateSingleSendCall(sendCalls: SendCalls, dappChainId: Hex) {
  * @param keyringType - The type of keyring associated with the account.
  */
 function validateSendCalls(
-  sendCalls: SendCalls,
+  sendCalls: SendCallsPayload,
   dappChainId: Hex,
   dismissSmartAccountSuggestionEnabled: boolean,
   chainBatchSupport: IsAtomicBatchSupportedResultEntry | undefined,
@@ -323,7 +326,7 @@ function validateSendCalls(
  * @param sendCalls - The sendCalls request to validate.
  * @throws JsonRpcError if the version is not supported.
  */
-function validateSendCallsVersion(sendCalls: SendCalls) {
+function validateSendCallsVersion(sendCalls: SendCallsPayload) {
   const { version } = sendCalls;
 
   if (version !== VERSION) {
@@ -340,7 +343,7 @@ function validateSendCallsVersion(sendCalls: SendCalls) {
  * @param dappChainId - The chain ID that the dApp is connected to
  * @throws JsonRpcError if the chain IDs don't match
  */
-function validateDappChainId(sendCalls: SendCalls, dappChainId: Hex) {
+function validateDappChainId(sendCalls: SendCallsPayload, dappChainId: Hex) {
   const { chainId: requestChainId } = sendCalls;
 
   if (
@@ -362,7 +365,7 @@ function validateDappChainId(sendCalls: SendCalls, dappChainId: Hex) {
  * @throws JsonRpcError if the chain ID doesn't match or EIP-7702 is not supported
  */
 function validateSendCallsChainId(
-  sendCalls: SendCalls,
+  sendCalls: SendCallsPayload,
   dappChainId: Hex,
   chainBatchSupport: IsAtomicBatchSupportedResultEntry | undefined,
 ) {
@@ -381,7 +384,7 @@ function validateSendCallsChainId(
  * @param sendCalls - The sendCalls request to validate.
  * @throws JsonRpcError if unsupported non-optional capabilities are requested.
  */
-function validateCapabilities(sendCalls: SendCalls) {
+function validateCapabilities(sendCalls: SendCallsPayload) {
   const { calls, capabilities } = sendCalls;
 
   const requiredTopLevelCapabilities = Object.keys(capabilities ?? {}).filter(