+++ packages/eip-5792-middleware/src/index.ts
@@ -8,24 +8,20 @@ export {
   getCapabilities,
   type GetCapabilitiesHooks,
 } from './hooks/getCapabilities';
-export { walletSendCalls } from './methods/wallet-send-calls';
-export { walletGetCallsStatus } from './methods/wallet-get-calls-status';
-export { walletGetCapabilities } from './methods/wallet-get-capabilities';
+export { walletSendCalls } from './methods/wallet_sendCalls';
+export { walletGetCallsStatus } from './methods/wallet_getCallsStatus';
+export { walletGetCapabilities } from './methods/wallet_getCapabilities';
 export type { EIP5792Messenger } from './types';
 
 export type {
   GetCallsStatusHook,
   GetCallsStatusParams,
   GetCallsStatusResult,
-} from './methods/wallet-get-calls-status';
-export type {
   GetCapabilitiesHook,
   GetCapabilitiesParams,
   GetCapabilitiesResult,
-} from './methods/wallet-get-capabilities';
-export type {
   ProcessSendCallsHook,
-  SendCalls,
+  SendCallsPayload as SendCalls,
   SendCallsParams,
   SendCallsResult,
-} from './methods/wallet-send-calls';
+} from './types';