+++ packages/eip-5792-middleware/src/methods/wallet_getCapabilities.ts
@@ -1,29 +1,9 @@
 import { rpcErrors } from '@metamask/rpc-errors';
-import type { Infer } from '@metamask/superstruct';
-import { array, optional, tuple } from '@metamask/superstruct';
-import type {
-  Hex,
-  Json,
-  JsonRpcRequest,
-  PendingJsonRpcResponse,
-} from '@metamask/utils';
-import { StrictHexStruct, HexChecksumAddressStruct } from '@metamask/utils';
+import type { JsonRpcRequest, PendingJsonRpcResponse } from '@metamask/utils';
 
+import { GetCapabilitiesStruct } from '../constants';
+import type { GetCapabilitiesHook } from '../types';
 import { validateAndNormalizeKeyholder, validateParams } from '../utils';
-
-const GetCapabilitiesStruct = tuple([
-  HexChecksumAddressStruct,
-  optional(array(StrictHexStruct)),
-]);
-
-export type GetCapabilitiesParams = Infer<typeof GetCapabilitiesStruct>;
-export type GetCapabilitiesResult = Record<Hex, Record<string, Json>>;
-
-export type GetCapabilitiesHook = (
-  address: GetCapabilitiesParams[0],
-  chainIds: GetCapabilitiesParams[1],
-  req: JsonRpcRequest,
-) => Promise<GetCapabilitiesResult>;
 
 /**
  *