+++ packages/eip-5792-middleware/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add EIP-5792 RPC method handler middlewares ([#6477](https://github.com/MetaMask/core/pull/6477))
+- Add and export EIP-5792 RPC method handler middlewares and utility types ([#6477](https://github.com/MetaMask/core/pull/6477))
 
 ## [1.0.0]